+++ kwarray/util_torch.py
@@ -40,7 +40,7 @@ def one_hot_embedding(labels, num_classes, dim=1):
         >>> t = one_hot_embedding(labels, num_classes)
         >>> assert all(row[y] == 1 for row, y in zip(t.numpy(), labels.numpy()))
         >>> import ubelt as ub
-        >>> print(ub.repr2(t.numpy().tolist()))
+        >>> print(ub.urepr(t.numpy().tolist()))
         [
             [1.0, 0.0, 0.0, 0.0, 0.0],
             [1.0, 0.0, 0.0, 0.0, 0.0],