+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Allow `None` as `indices` in `kwarray.Sticher.add`
 * Change `nan_behavior` to `nan_policy` in `kwarray.RunningStats`
 * Avoiding torch inputs when possible. `util_slider` now imports torch locally if it needs to.
+* Updated `ub.repr2` to `ub.urepr`.
 
 ### Fixed
 * `robust_normalize` now preserves masked arrays.